Don't create ACL objects when defaultAcl mode is OVERRIDE
If defaultAcl.mode is set to OVERRIDE, don't create the virtual
containers for the file system ACLs. If the file system content has
been indexed previously, the ACL virtual containers will be deleted
when using OVERRIDE. The deletion will happen as the virtual
containers are polled for re-indexing, not immediately.

Bug: 141337637